requirements.txt
python-can>=4.3.0
matplotlib>=3.8.0
rich>=13.7.0
numpy>=1.26.0
msgpack>=1.0.0
